MATLAB TM Bounding Box
The previous algorithm was based on morphological operations which were
not general enough under different lighting and noise conditions.

The new algorithm expands on the idea of removing the belt influence by
computing the box based on the upper half of the side view. It assumes a
good background subtraction, but does not rely on a contiguous area. It
simply finds the first point to the right where the sum of a thresholded
image exceeds a certain level.